Added instructions and fixed some typos
Someone could think that instructions are useless since everyone on GitHub will know how to copy a repository, but what about a casual user or someone not familiar/with zero experience with the platform? I mean, not every single person on the planet knows how to use GitHub or knows it even exists.